Track a stable version of rocket-chip repo with 1.7 ISA
This allows loading RISC-V Linux with spike (local dev) or fesvr (board) with the command "spike +disk=root.bin bbl vmlinux".

The riscv-tools (and thus the master branch of rocket-chip) that folllow the 1.9 ISA have changed the riscv-tools/riscv-pk's bbl (bootloader) and the above syntax doesn't work since bbl needs to be built with having a Linux kernel payload (vmlinux). However even after building such a bbl it doesn't work as of 16/7/2016.